backend/requirements-directml.txt
torch
transformers>=4.21.0
onnxruntime-directml
librosa
numpy
soundfile